added reference to Xie et al 2012 to Background to add an example of stragglers from network competition
I can suggest “https://ieeexplore.ieee.org/document/6468446” on
the straggler issue as a reference. It is an old paper we wrote
for SC’12, and you can find others as well, but since this is the
main theme of your paper, including some more references on the
subject might not hurt.