more details on (qualitative) straggler definition
We don't have a precise definition of stragglers, partly because our stragglers do not quite
seem to fit the definition of stragglers in the MapReduce literature. However, we also don't
need a precise definition because we only use straggler identification in a qualitative way
to improve our performance. Given the phenomenological nature of the paper, the following
should be sufficient

- more detailed review of straggler definition
- made clear origins in MapReduce
- in Results: made clear that we define stragglers in a very qualitative fashion but that
  this is sufficient for our purposes