removed Fig A4 and add t_comm to Fig 6
Fig A4 (subfiling on SuperMIC with MPI and GA) was removed as the GA data are not
included anymore and the MPI subfiling it was mostly duplication of
data shown in (now) Fig 6. The removal of the GA graphs allowed us to add the
communication time to Fig 6c for Comet and SuperMIC (which was originally
shown in Fig A4c) without compromising the clarity of the presentation.

Renamed figure directories to match draft.